[huff0] Add x86 specialisation of Decode4X
BenchmarkDecompress4XNoTable/gettysburg-32           593.83       681.84       1.15x
BenchmarkDecompress4XNoTable/twain-32                491.42       680.16       1.38x
BenchmarkDecompress4XNoTable/pngdata.001-32          718.28       870.23       1.21x
